experiments, exploration

+++ utils/A_star/A_star.cpp
@@ -16,7 +16,7 @@ AStar::~AStar() {}
 double AStar::pathLength(const vector<Eigen::Vector3d>& path) {
   double length = 0.0;
   if (path.size() < 2) return length;
-  for (int i = 0; i < path.size() - 1; ++i) length += (path[i + 1] - path[i]).norm();
+  for (size_t i = 0; i < path.size() - 1; ++i) length += (path[i + 1] - path[i]).norm();
   return length;
 }
